requirement.txt
numpy
pandas
matplotlib
scikit-image
scipy
nibabel
antspyx
dipy
fslpy
Cython
SimpleITK==2.0.2
